requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml==5.2.2
